reflow/ec2cluster: Backoff SpotProbeDepth upon InstanceLimitExceeded errors
Summary:
Since we use a starting SpotProbeDepth of 20, often times
we'll hit an InstanceLimitExceeded due to limits on the AWS accounts.

That can cause use to use potentially more expensive instances while
a cheaper instance may very well be available.

This chance backs off SpotProbeDepth and retries immediately when
such an error is encountered.

Test Plan: Tested with a custom config limiting to some instance types with <20 instance limits.

Reviewers: marius, pgopal, jcharumilind

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D36684

fbshipit-source-id: c9fed67