reflow/alloc: allow ref counting of alloc objects
Summary:
This diff changes the executor API to explicitly let the alloc know the
liveset for an exec using the Load/Unload calls. The executor reference
counts the objects and is free to delete the objects with ref count 0.
This diff removes any object with zero reference count.

Test Plan:
units tests added
ran a test which shared lot of data between execs in multiple stages. Ran with strict collection and did not run into any issues.

Reviewers: jcharumilind, marius, smahadevan

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D37271

fbshipit-source-id: c9e1335