reflow/alloc: change executor api to allow alloc garbage collection
Summary:
This diff changes the executor API to explicitly let the alloc know the
liveset for an exec using the Load/Unload calls. The executor reference
counts the objects and is free to delete the objects with ref count 0.
This diff removes any object with zero reference count.

The previous change incorrectly incremented the ref count without
checking dead objects, which could result in incrementing the ref count
of an object that is marked for deletion. In the current diff,
Executor.refCountObject does that and is used in all places where we
increment the ref count.

Test Plan: Added unit test and ran a test program

Reviewers: marius, jcharumilind, smahadevan, dnicolaou

Reviewed By: marius

Maniphest Tasks: T28590

Differential Revision: https://phabricator.grailbio.com/D38696

fbshipit-source-id: 99bb83a