reflowbootstrap: Refactor grail and OSS bootstrap to share code
Summary:
Refactor the bootstrap server code and share most functionality
between the grail and open source bootstrap binaries.

Haven't updated the docker image for both Grail and OSS with this change,
because for the most part this change is a no-op, but I have tested it.

Test Plan:
For grail, built an image with tag `reflowbootstrap_test` and verified
that a simple reflow program `reflow/remote_test` runs correctly.

For OSS, built my own docker image (`smahadevan/reflowbootstrap_test`)
and tested on my personal AWS account.

Reviewers: dnicolaou, jcharumilind, pgopal, marius

Reviewed By: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D37354

fbshipit-source-id: 3f23627