reflow/s3blob: Use a generous timeout for unknown file size
Summary:
When the file size is unknown, use a 5m timeout instead of 1m.

At least one use case where this happens is when we are uploading
the reflow image to the repo for later installing on a reflowlet,
specifically on a slow internet connection.
(For example on my home network it takes >1m for 41M file)

Though we have a large timeout in `instance.go` (5m), we fail
to upload due to this deeper timeout.

Reviewers: marius, jcharumilind, pgopal

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D37280

fbshipit-source-id: fc80e8c