reflow/executor: Round time to the nearest second before calculating runtime
Summary:
Currently, `ExecInspect.Runtime()` will said the the container endtime as the current time if the end time is before the start time. This is problematic because the current time is not necessarily similar to the end time whatsoever. For example, the following ExecInspect is for an exec that was run on 2019-05-30: `sha256:67f6bd512895edf05fdbd0331a658286726235e079a83c613f6e03e679c66fa4`. The ExecInspect contains the following information from docker:

{F958207}

Upon close inspection, it can be seen that `FinishedAt` is around 0.03 seconds before `StartedAt`. If I were to call `Runtime()` on this ExecInspect today, 2019-10-04, the calculated runtime would be around 4.15 months, which makes no sense. In order to remedy this, D37291 makes `ExecInspect.Runtime()` assume that time is accurate to the nearest second, and rounds the start and end times to the nearest second before calculating runtime.

Test Plan: Scan `reflow-meta`, and get runtime data. Ensure that all runtimes are calculated to the nearest second.

Reviewers: jcharumilind, pgopal, marius, smahadevan

Reviewed By: smahadevan

Tags: #reflow

Differential Revision: https://phabricator.grailbio.com/D37291

fbshipit-source-id: c45d86b