reflow/bootstrap: Add release scripts for Grail/OSS reflow bootstrap binaries
Summary:
Add scripts that build the reflow bootstrap binary (Grail and OSS versions)
and push them to a public S3 bucket.

In a subsequent revision, we will change reflow to fetch the bootstrap binary
from these S3 locations to bootstrap the reflowlet running on an EC2 instance.

Test Plan: Will be tested in a subsequent revision which uses these paths.

Reviewers: marius, bbentson, jcharumilind, pgopal

Reviewed By: bbentson

Maniphest Tasks: T27890

Differential Revision: https://phabricator.grailbio.com/D37547

fbshipit-source-id: b42b680