reflow: Reflow version in config should always come from the binary
Summary:
The reflow version used in the configuration should always be
set from the binary regardless of what's specified (if any) in a provided config.

Also, avoid marshaling the version upon `reflow config -marshal`.

Test Plan:
Previously, if we pass a config created by reflowv1 (ie, `reflowv1 config -marshal`)
to reflowv2 (using `-config` flag), the run will stall because it'll keep creating
reflowlets that run reflowv2 but the EC2 cluster will be looking for reflowv1.

With this change, the following will work as expected.
```
> reflowv1 config -marshal > /tmp/config
> reflowv2 -config /tmp/config run reflow/remote_test.rf
```

Reviewers: marius, pgopal, jcharumilind

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D37587

fbshipit-source-id: e1ede1c