reflow/s3blob: Use a smaller limit for HeadObject calls
Summary:
Reflow's initial evaluation of a module involves looking up flow's result
in the cache and validating them.  Validating caches results requires refreshing
their assertions which results in tons of S3 HeadObject calls.

We seem to be seeing a big latency and lot of deadlines when making a massive
number of HeadObject calls in parallel.  Until we figure out and fix the
underlying problem, this limits the number of concurrent calls.

Reviewers: jcharumilind, pgopal, dnicolaou

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D39307

fbshipit-source-id: 556ff54