reflow/local: Add system OOM checking for dockerExecs
Summary:
Currently, reflow cannot detect if a dockerExec was killed by the system OOM killer. D37942 adds methods to search `/dev/kmsg` for OOMs due to a process being killed by the OOM Killer or if a process exceeded its cgroup's memory limit. This allows for a `dockerExec` to be marked as OOMKilled if the docker container fails to detect that it was OOMKilled.

This works as follows:
1. `local.Executor` reads `/dev/kmsg` in a goroutine and maps all OOMed PIDs to their timestamps. The goroutine is excited via the context. This is done because the scanner will never finish because `/dev/kmsg` does not end with an `EOF`.
2. `local.dockerExec` checks the OOM map after the docker container is exited to see if the container OOMed.
3. The map is contained in a struct called `local.oom` along with a `sync.Mutex` to support concurrent read/writes.

Test Plan: Unit tests

Reviewers: jcharumilind, pgopal, marius, smahadevan

Reviewed By: jcharumilind, marius, smahadevan

Tags: #reflow

Differential Revision: https://phabricator.grailbio.com/D37942

fbshipit-source-id: 35ec5a2