reflow/syntax: Do not display ignored parameters in reflow doc
Summary:
Currently, running `reflow doc` on a `.rfx` file, the following will occur:

```
val aligner string = "siege"
    aligner is either "siege" or "bwameth".
val _ string = "siege"
    aligner is either "siege" or "bwameth".
```

This is due to the following:

When a reflow bundle is evaluated, args are injected into its environment. In `module.Params()`, `DeclAssign`s are double counted in the sense that they are added to the environment with their `Ident` and with an ignore batter (`_`). D38059 prevents ignore patterns from being returned by `module.Params()`.

Reviewers: jcharumilind, pgopal, marius, smahadevan

Reviewed By: jcharumilind, marius

Subscribers: escott

Tags: #reflow

Differential Revision: https://phabricator.grailbio.com/D38059

fbshipit-source-id: ea17e18