reflow/eval: Limit concurrent (un)marshalling of cached filesets
Summary:
Json encoding/decoding is a very intensive process especially for
large filesets which are encountered upon cache lookups (and writes).
It also requires the whole json to be held in memory and hence parsing
too many of these concurrently can result in significant memory pressure
(and OOMs even on large-ish adhocs).

This is a band-aid fix while the long-term solution would be to provide
a `Fileset` implemenation optimized for fast encoding/decoding as well as
compact in-memory representation.

Test Plan:
Tested on a large pipeline which kept failing with OOMs before this
change and is able to run with this change.

Reviewers: marius, jcharumilind, pgopal

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D38594

fbshipit-source-id: 7955ef5