reflow/repository: use separate errgroups for transferers and waiters
Summary:
Currently we are possibly masking an error somewhere in the transfer
manager's transfer() method. We dedup transfers by having only one
transfer of a specific file from src->dest. A dup transfer request simply waits
for the first transfer request to complete. Since the waiting goroutine
waits for either the transfer.C or ctx.Done() channels to return,
there is a chance for a race and possibly masking the actual error.

If the original transfer routine errored out and returns on the
transfer.C channel before the group's error can be canceled, then we get
the original error. But if the group's context cancelation beats the
tranfer.C channel, then we will see a context cancellation error instead
of the actual error. This change only breaks up these error groups and
provides their own context.

Reviewers: jcharumilind, smahadevan, dnicolaou

Reviewed By: smahadevan

Differential Revision: https://phabricator.grailbio.com/D39360

fbshipit-source-id: 7b08174