reflow/integration: Move cache assertion tests to be under grail.com
Summary:
These tests rely on the existence of an S3 repository, assoc and
a specific S3 bucket (where the test files are put) and hence are not
easily runnable in the OSS distribution.  It doesn't make sense to
mock those infrastructure pieces either since this is an integration test.

Test Plan: Unit/Integration tests.

Reviewers: sbagaria, pgopal, jcharumilind, dnicolaou

Reviewed By: sbagaria, pgopal, dnicolaou

Differential Revision: https://phabricator.grailbio.com/D39630

fbshipit-source-id: a39eb3d