reflow/tool: Always get a repo/assoc
Summary:
We always try to get a Repository and Assoc but only fail on error
if we are setup to use cache.

This makes sure that we'll get an Assoc/Repository if available even
if cache is off, because we seem to need it for TaskDB.

Test Plan:
Unit/Integration tests.

Also tested with the following changes in `reflow/modules/BUILD`:
```
 reflow_module_test(
     "internal/bio_count",
+    reflow_params = ["-log=debug", "-cache=off"],
     images = ["bio-count"],
 )
```
and ran the test:
```
> bazel test //reflow/modules:internal/bio_count_test
//reflow/modules:internal/bio_count_test                                 PASSED in 115.0s
```

Reviewers: jcharumilind, pgopal, dnicolaou

Reviewed By: dnicolaou

Tags: PHID-PROJ-p5ztqpc3gwzgnkywvwge

Differential Revision: https://phabricator.grailbio.com/D39818

fbshipit-source-id: 0678594