reflow/eval: resolve unresolved files before comparing them
In the scheduler mode, we delay interning files to when execs need them.
This means if we try to compare two files, one that is a result of an
exec and one that isn't dependent on an exec, we end up with two files
where one is fully resolved and the other isn't.

This diff changes the behavior to panic if we ever compare a resolved
file and an unresovled one.

We now force file resolution when we try to compare two files where one
is resolved and the other isn't.