reflow/local: Allow execs to have access to host docker daemon
Summary:
If an exec needs access to the host's docker daemon, it can simply
include the suffix `$docker` to the image and the docker sock path
will be bound when the container is created.

Ref T30751

Test Plan: This feature was requested by @spolakh and I assume he will be testing it.

Reviewers: pgopal

Reviewed By: pgopal

Subscribers: jcharumilind, dnicolaou, spolakh

Tags: #reflow

Maniphest Tasks: T30751

Differential Revision: https://phabricator.grailbio.com/D40496

fbshipit-source-id: 513d448