Revert "reflow/flow: Set minimum CPU limit to 0.1"
Summary:
This reverts commit 42539d32f3be8cfa147c50d8130877dcf0b994e2.

Floating point arithmetic causes us to lose precision. This results in a
cpu value of 31.996 or something similar when the actual cpu is 32. This ends
up us not scheduling an exec which needs all 32 cpus. This issue wouldn't be
noticeable in reflow's cluster mode as those resources could be used by exec's
needing lesser cpu.

This is an issue in production which runs local mode and where execs are tightly packed.
This issue was observed on a r3.8xlarge which has 32 cpus. We end up in a state where
we are running nothing and we have 2 runnable execs that both require 32 cpus. Since we
have only 31.997 (due to floating point math), we end up panicing with "scheduler stuck".

Test Plan: Reverted this change and ran in dev cluster and the pipeline completes

Reviewers: dnicolaou, jcharumilind

Reviewed By: jcharumilind

Subscribers: cbain, ghorrell, pbohannon, jcharumilind, smahadevan, dnicolaou

Maniphest Tasks: T31702

Differential Revision: https://phabricator.grailbio.com/D41484

fbshipit-source-id: 44883c9