reflow/s3blob: fix bug from D40195 where we ignored an error
Summary:
We dropped an error when the duration of the head object call took
longer than the limit. This caused us to incorrectly return that the
object exists when it did not.

This is the source of multiple users running into cache write missing
file errors.

Test Plan: Ran a test program and it correctly returns the error now.

Reviewers: jcharumilind

Reviewed By: jcharumilind

Subscribers: smahadevan, dnicolaou

Maniphest Tasks: T32071

Differential Revision: https://phabricator.grailbio.com/D41755

fbshipit-source-id: a7086c8