reflow: use Predictor during flow evaluation
Summary:
Use the `Predictor` in the flow evaluator to optimize resource usage.
Reflow users will specify if they want to use the `Predictor` with the `-pred` flag.
`-pred` will be set to `false` by default for now. `-pred` will be set to `true` by default
once we are confident that it will not interfere with existing workflows. This will require
testing with existing research pipelines.

Note that in order for the `Predictor` to be used both `-sched` and `-pred` must be `true`.
This is because the `Predictor` will only work with the scheduler.

D42723 is stacked on top of D42706.

Test Plan:
Manually ran the following reflow program:

```
val Main = exec(image := "ubuntu", mem := 10*GiB) (out file) {"
	echo hello world >>{{out}}
"}
```

This resulted in the following log message:

```
reflow: task 2e9f8ad4: modifying resources from {mem:10.0GiB cpu:1 disk:0B} to {mem:500.0MiB cpu:1 disk:0B}
```

Additionally, I checked the task's `ExecInspect` and found that its allocated memory was `500*MiB`:

```
    "Resources": {
      "cpu": 1,
      "disk": 0,
      "mem": 524288000
    }
```

Additionally, verified that the `Predictor` will not turn on on my mac even if `-pred=true` and that it will turn on on my adhoc if `-pred=true`.

Reviewers: jcharumilind, smahadevan

Reviewed By: jcharumilind, smahadevan

Subscribers: pgopal

Tags: #reflow

Maniphest Tasks: T30968

Differential Revision: https://phabricator.grailbio.com/D42723

fbshipit-source-id: 1b75451