Compute a proper flow digest for switch continuations (#119)
Compute a proper flow digest for switch expression evaluation continuations. The existing implementation was very incomplete and results in equivalent digests for non-equivalent expressions. The digest now includes the entire remainder of the switch expression to be evaluated.

Unit tests are added to verify.

This also contains a bit of refactoring to eliminate stutter ("evalSwitch.eval*").