pipeline/module: Force intern of missing input files
Summary:
Reflow always stages a (previously) cached object if it is needed
by an exec as input.  These grail modules support referencing an
S3 file but they are not execs.  So Reflow is otherwise unaware
that these objects (if cached) need to be staged.

In this change, we add a check to see if the object is missing
from the repository and force an intern.
Additionally, forced interns weren't always working and that's
also been fixed in this revision.

Test Plan:
This issue was found in a production pipeline and reliably reproduced
locally (See https://phabricator.grailbio.com/T36095#1050779)
Confirmed that this change makes that pipeline avoid the issue and proceed further.

Reviewers: pgopal

Reviewed By: pgopal

Subscribers: jcharumilind, dnicolaou, kshashidhar

Maniphest Tasks: T36095

Differential Revision: https://phabricator.grailbio.com/D45057

fbshipit-source-id: 879ec1a