reflow/filerepo: Add missing integrity check
Summary:
Reflow transfers objects from between different repositories
using the most optimal mechanism depending on the src/dst repo types.

If the source is S3, we do a multi-part download to make things go faster.
In that case, we should still make sure to do an integrity check.
That is, confirm that the digest of the object bytes transferred matches
the one we expected for that object.

Also, we add retries for all failed transfers if the reason was transient
or an integrity failure.

Test Plan: Unit tests and integration test (added in a subsequent revision)

Reviewers: pgopal

Reviewed By: pgopal

Subscribers: jcharumilind, dnicolaou

Maniphest Tasks: T36183

Differential Revision: https://phabricator.grailbio.com/D45272

fbshipit-source-id: 1db0a9d