reflow/sched: submit should copy the task slice
Summary:
Currently we pass the slice from eval to sched over a chan.
This means that eval is modifying the slice, while sched is working on
the task slice. This hasn't popped up until now. but when we run large
sets of runs using a single scheduler, we start seeing the race and the
program panics as it sees old tasks that have already been submitted.

Test Plan: ran a giant batch program that ran into this bug. It goes away with this fix.

Reviewers: smahadevan

Reviewed By: smahadevan

Subscribers: jcharumilind, dnicolaou, spolakh

Maniphest Tasks: T36539

Differential Revision: https://phabricator.grailbio.com/D45392

fbshipit-source-id: cdd0e35