reflow/executor: On Put, treat completed but failed execs as new
Summary:
The `Executor` saves and retains the state of all execs that have
been `Put` and even restores this state from `Dir` upon a restart.
In practice, this means that often failed execs cannot be retried
on the same `Executor`.

Instead, when `Put` is called on an `Executor` for the same exec,
and if the exec is already complete but ended up with an error,
then we allow the exec to be created again.

Test Plan:
Unit tests.

Ran a sample program (which sleeps for 5s and then fails) with old `reflow` binary and it returns re-attempts immediately:
```
2020/05/20 14:45:48  ->  sleep.Main   5a0fce3c exec   exec ..f42b2d6d81e3d743e2ab571f55bcd5df788cc8 sleep 5.echo "Hi!" > {{out}} && false
2020/05/20 14:45:48  <-  sleep.Main   5a0fce3c err    exec 5s 0B
```

With this change, the exec is re-run (ie, the program sleeps again):
```
2020/05/20 14:47:41  ->  sleep.Main   5a0fce3c exec   exec ..f42b2d6d81e3d743e2ab571f55bcd5df788cc8 sleep 5.echo "Hi!" > {{out}} && false
2020/05/20 14:47:49  <-  sleep.Main   5a0fce3c err    exec 5s 0B
```

Reviewers: pgopal

Reviewed By: pgopal

Subscribers: jcharumilind, dnicolaou

Tags: #reflow

Maniphest Tasks: T36642

Differential Revision: https://phabricator.grailbio.com/D45463

fbshipit-source-id: 84a5bbe