reflow/infra: add NonEC2Ok field to PredictorConfig
Summary:
Add `NonEC2Ok ` to `infra.PredictorConfig`
so that users have the option of running the `Predictor`
non-ec2-instance machines. By default, `NonEC2Ok ` will be
set to `false` so that reflow will only use the `Predictor` if
reflow is running on an ec2 instance.

For testing purposes, however, it is useful to run the `Predictor`
on non-ec2-instance machines. This is appropriate for small integration tests
that will only run on a few execs and will only fetch small
amounts of data because these tests will not require many
calls to `taskdb` or `repo`.

Additionally, a test predictor config is registered in the infra package
with `NonEC2Ok` set to `true` and `MinData` set to `1`
for use in the integration test added in D45451.

Test Plan:
`bazel run //go/src/grail.com/reflow/integration:remote_test`
passes using these changes in D45451.

Reviewers: smahadevan

Reviewed By: smahadevan

Subscribers: jcharumilind, pgopal

Tags: #reflow

Maniphest Tasks: T30968, T36474

Differential Revision: https://phabricator.grailbio.com/D45605

fbshipit-source-id: f3e8a14