reflow/flow: make all ExecConfig values from the Flow be copies of the Flow values
Summary:
Currently, flow values are copied directly
into the `ExecConfig`. This is dangerous
for maps and slices (`f.Resources`, `f.OutputIsDir`)
because if these `ExecConfig.Reserved` or
`ExecConfig.OutputIsDir` are modified, the
respective flow values are changed as well.

This problematic in the case of `ExecConfig.Resources`
because the resources may be modified if `Predictor`
is in use. This causes the bug outlined in T36816. In
T36816, an exec's resource usage was dropped from
`400GiB` to `14GiB`. Due to the `ExecConfig.Resources`
being the same slice as `f.Reserved`, modifying the
`ExecConfig` modified the reserved resources. This lost
track of the flow's original resources. When the lack of
memory caused an OOM, reflow calculated the new
memory requirements off of the bad prediction, `14GiB`,
instead of the user-defined resources, `400GiB`.

By making sure the `ExecConfig` receives a **copy** of all
the `Flow`'s values, bugs like the one in T36816 can be
avoided.

Note that this does not affect the relationship between
`Flow.Resources`, `Flow.Reserved`, and `Task.Config.Resources`.
The relationship between the user-specified resources, the
system-adjusted resources and the flow's reserved resources
needs to be revised in a follow-up diff (see T36985).

Test Plan:
Re-ran the job in T36816. The run did not fail due
to an OOM because the following occurred:

```
2020/05/21 19:48:26 task 7b49f84c: modifying resources from {mem:400.0GiB cpu:30 disk:0B} to {mem:16.3GiB cpu:30 disk:0B}
...
2020/05/21 19:51:45 flow 98c27fda: OOM: re-submitting task 91e3118d with 600.0GiB of memory (1/3)
```

I manually checked the dynamodb backend to ensure that task `7b49f84c`
corresponds to flow `98c27fda`.  Additionally, I also added the flow id
to the `modifying resources` logging message to make future debugging
easier.

Reviewers: smahadevan

Reviewed By: smahadevan

Subscribers: jcharumilind, pgopal

Tags: #reflow

Maniphest Tasks: T36816

Differential Revision: https://phabricator.grailbio.com/D45634

fbshipit-source-id: c137432