reflow/ec2cluster: fix the resource allocation for wide requests
Summary:
This change fixes a bug associated with wide requests. It also
simplifies the code and adds tests.
The allocation is not optimal and will be fixed in a future
change.

The bug is on lines 531-539 where we ignore the width of that request. If a non wide request is followed by a wide request, then in some cases, we ignore the width of the wide request.

Test Plan: unit tests

Reviewers: smahadevan

Reviewed By: smahadevan

Subscribers: kshashidhar, jcharumilind, dnicolaou, spolakh

Maniphest Tasks: T19475

Differential Revision: https://phabricator.grailbio.com/D45680

fbshipit-source-id: 73460ef