reflow/sched: restore task filesets on failure
Summary:
Task failures can result in tasks being assigned to new allocs. Once we
load input filesets, we mark them as resolved for this alloc.
But if this task fails and gets reassigned to a new alloc, those "resolved"
files need to be resolved again for the new alloc.

Test Plan: unit test

Reviewers: smahadevan, dnicolaou

Reviewed By: dnicolaou

Subscribers: pboyapalli, jcharumilind, dnicolaou

Maniphest Tasks: T38084

Differential Revision: https://phabricator.grailbio.com/D46484

fbshipit-source-id: 4d69728