reflow/ec2cluster: Improve EC2 cluster state synchronization
Summary:
Reflow's EC2 cluster state is defined as the set of instances
returned by AWS API that have some set of tags (specified by `QueryTags()`)

Despite extensive searching, I'm unable to find how much (if any)
propagation delay exists between when an instance is tagged and it being
returned by AWS APIs when queried with those tags.

Currently, when we start a new instance, as soon as its up and tagged,
we consider it available and move on to service other requests.
However, its likely that the instance may not be part of the cluster (yet)
despite a forced synchronization of the cluster state.

This revision addresses these issues:
- Ensure newly minted EC2 instances are indeed part of the cluster before counting them as no longer `pending`.  This helps avoid starting more than the total allowed instances (in the config).
- The above also prevents duplicate instance allocation requests for the same underlying requirements.  (Unless the resources got swooped up by a different set of requirements, in which case we really do need more)
- Prevent the cluster expansion go-routine (`loop()`) from waiting for a cluster state synchronization.  Instead, this is now done in go-routines dedicated for launching each EC2 instance (`launch()`).

Test Plan:
Ran a sample program which would require upto 20 instances, based on the instance types specified in the config and if the max instances were >=20.
Without this change, if max instances was set to 5, we could see 5+ instances running.
With this change, upon multiple tries we didn't have more than 5 instances running.

Reviewers: pgopal, pboyapalli

Reviewed By: pboyapalli

Subscribers: jcharumilind, pgopal, dnicolaou, pboyapalli

Tags: #reflow

Differential Revision: https://phabricator.grailbio.com/D47236

fbshipit-source-id: fa1cd89