reflow/scheduler: Tasks failing on lost allocs must be retried
Summary:
Tasks can fail on allocs with certain type of errors which
indicate a lost alloc.  In those cases, we should correctly
consider the task as lost and allow it to be re-attempted
on a different alloc.

Previously, the main issue was the check `err == ctx.Err()`
which isn't true because the `err` is usually `errors.Error`
type and isn't equal to the context error.

Test Plan:
Unit tests.
Integration tests will be added in a follow-up revision.

Reviewers: pgopal

Reviewed By: pgopal

Subscribers: dnicolaou, spolakh, pboyapalli

Maniphest Tasks: T38718

Differential Revision: https://phabricator.grailbio.com/D48070

fbshipit-source-id: 76601cc