reflow/predictor: Support prediction of task duration
Summary:
Add the ability to predict exec duration for tasks submitted to the scheduler.

This could potentially be used by the scheduler when allocating resources.

Ideally, we should be storing and saving the total time taken for a task
(ie, including load/unload times, etc), but this computes the time taken
by the exec which is the least predictable (from reflow's perspective).
The time taken for data transfer, etc is relatively easy to compute based
on fileset size.

Test Plan: Unit and integration tests.

Reviewers: pgopal, dnicolaou

Reviewed By: dnicolaou

Subscribers: pboyapalli

Differential Revision: https://phabricator.grailbio.com/D48230

fbshipit-source-id: 5c1c3ce