reflow/tool: dont print result if eval resulted in an error
Summary:
This was causing a local run to panic with an interface conversion
error. If the evaluation resulted in an error, we shouldn't try to
interpret the result value.

Test Plan: ran a test program that used to fail. now it doesnt panic and it reports the assertion failure.

Reviewers: smahadevan

Reviewed By: smahadevan

Subscribers: dnicolaou, pboyapalli

Maniphest Tasks: T40209

Differential Revision: https://phabricator.grailbio.com/D48212

fbshipit-source-id: 4acf760