microserv/template: revert update to sqlboiler v4
Summary:
This reverts commit 2e9f7497d723dba8c269455912415e4579ad10f9.

Though this was fine for the template microservice, it doesn't have
support for CamelCase schemas, so it broke a few of the other services.

Since camelcase is the standard for db schemas at grail, choosing to
revert this.

There's a now-abandoned PR for adding support for this to sqlboiler that
may be worth looking into eventually https://github.com/volatiletech/sqlboiler/issues/557

Also, as a bonus: cleaned up a few git merge artifacts.

Test Plan:
This simply reverts a previous commit. That said, I
regenerated the models for the shipment and reference service and they
build correctly.

Reviewers: achew, adevarmanai, O3 third-party

Reviewed By: achew, O3 third-party

Subscribers: joshnewman, pgopal, smahadevan, dnicolaou, pboyapalli

Maniphest Tasks: T38880

Differential Revision: https://phabricator.grailbio.com/D48464

fbshipit-source-id: 67d0687