reflow/pool: Consider multiple picks when attempting to allocate
Summary:
When attempting to allocate from a pool (typically a mux across say P pools),
instead of considering just one "best" pick (on each of 6 tries),
we consider upto N picks (N set to a default of 10) ordered by "best match".
Implementation uses a fixed size heap (so that the complexity is `P log N` instead of `P log P`).

This makes a big difference especially in the case where multiple allocation requests
come in with about the same requirements across a large pool with potential allocs of the same size.
This is a common pattern seen in production pipelines with `pi-evaluator`
and is simulated in the newly added `TestAllocateScale`.

For example, if the old `pick` is used, the newly added test `TestAllocateScale`
consistently fails with `30%` errors while the new `pickN` very rarely fails with `<1%` errors.

This means that in production `pi-evaluator` use-case, upto `30%` of the time we could fail to
allocate from the existing pool (even when allocs where available) and go for expanding the cluster
instead.  What's worse is that, a particular unlucky request could starve for a very long time.

Test Plan: Unit tests and integration tests.

Reviewers: pgopal, pboyapalli

Reviewed By: pgopal, pboyapalli

Subscribers: dnicolaou, ghorrell, spolakh, pboyapalli

Tags: #reflow

Maniphest Tasks: T40569

Differential Revision: https://phabricator.grailbio.com/D48947

fbshipit-source-id: ad80f21