reflow/syntax: digest exprexec in stdEvalK
If an exec has delayed dependencies, we need to digest them while computing the flow digest. Without this, execs with delayed dependencies can result in having same logical digests, which is incorrect. 

No tests yet. will add some. Also i feel like we should cover all the cases here that are relevant and panic for cases which are no supposed to use this (like compr).