reflow/eval: Add ctx to flow rendezvous mechanism in tests
Summary:
Many of these unit tests rely on a mechanism by which the evaluator
is expected to put a flow in the executor which waits for it to show up
and then updates its result to be whatever the test wants.

The synchronization can fail if any conditions setup by the test (and expected to happen)
didn't occur, which can lead to the test hanging forever.
This makes debugging tests very difficult.

Adding ctx to the rendezvous mechanism allows us to fail the test early
and also provides a bit more clue to where the issue might be.

Test Plan: Unit tests.

Reviewers: pboyapalli, pgopal

Reviewed By: pboyapalli, pgopal

Subscribers: pgopal, dnicolaou

Tags: #reflow

Differential Revision: https://phabricator.grailbio.com/D49396

fbshipit-source-id: e45b749