reflow/syntax: don't use saved exec image name in the exec flow
Summary:
Though uncommon, a single reflow program can use different images for
the same exec. As written before this diff, the last image usually gets
picked up as the image for all the exec instances, which is incorrect.
It is also important to note that this issue manifests when there is at least one delayed dependency to the exec, not otherwise.

Test Plan: unittest

Reviewers: smahadevan

Reviewed By: smahadevan

Subscribers: dnicolaou, pboyapalli

Maniphest Tasks: T42428

Differential Revision: https://phabricator.grailbio.com/D50694

fbshipit-source-id: 1be1cbc