reflow/trace: add `New` method for `LocalTracer`
Summary:
This diff adds a `New` method for `LocalTracer` which is intended for cases
where Reflow is used as a library and the user would like to specify the
path of the output trace file.

Depends on D50569

Test Plan: New and existing unit tests pass.

Reviewers: smahadevan, ghorrell

Reviewed By: smahadevan

Subscribers: dnicolaou

Tags: #reflow

Maniphest Tasks: T45480

Differential Revision: https://phabricator.grailbio.com/D52913

fbshipit-source-id: a83a8ce