reflow/tool: log reflow version with the exec logger
Summary:
When debugging, it is useful to know the version of Reflow used to start the
run. We [[ https://sg.eng.grail.com/grail/grail/-/blob/go/src/github.com/grailbio/reflow/tool/main.go#L316 | already log the version here ]],
but it is at the debug level and only gets logged to stderr in the user's terminal.

This diff uses the exec logger to log the version so that it is included in the `.execlog`.

Test Plan:
Built a binary with this change included and ran a sample program. Verified that
the result execlog had the reflow version.
```
$ bazel run //go/src/grail.com/cmd/releasereflow -- -dirty Scratch && /tmp/darwin/amd64/reflowScratch -log=debug -cache=off run -local ~/reflow_programs/sleep.rf
...
$ head -n 2 ~/.reflow/runs/a29d0fe902417069f3aa2ddda81aa5c6ac6bae2c20e8a3869fe54879052ffa44.execlog
2021/02/23 14:10:15 reflow version: reflowScratch (GRAIL, go1.13.15)
2021/02/23 14:10:15 run ID: a29d0fe9
```

Reviewers: smahadevan

Reviewed By: smahadevan

Subscribers: dnicolaou

Maniphest Tasks: T46941

Differential Revision: https://phabricator.grailbio.com/D55186

fbshipit-source-id: 19af8f0