reflow: fix reflow readme sample program
Summary:
This diff fixes the reflow sample program given in README.md. As with
D52984, the sample program was only requesting 1GB of memory, but it
needed more. The instance where reflow ran that exec didn’t have
sufficient memory so the program failed with a "Cannot Allocate
Memory" error. This fix updates the program so that it requests
the appropriate amount of memory and reflow places the exec on an
appropriate instance type.

Test Plan: Ran the sample program locally to completion.

Reviewers: pboyapalli, smahadevan

Reviewed By: pboyapalli

Subscribers: dnicolaou

Maniphest Tasks: T45535

Differential Revision: https://phabricator.grailbio.com/D55397

fbshipit-source-id: efd9dcf