reflow/taskdb: Add support for tracking task attempts
Summary:
Add support for tracking multiple attempts made to complete a task.
We add a new column `Attempt` to the taskdb which'll store the
attempt number for that task.  When the scheduler re-attempts the
task (typically done by changing its internal state to `TaskInit`),
the attempt number is incremented and thus helps distinguish the
subsequent task row in the taskdb.

This will particularly help with more accurate cost computation for
a reflow run (because we should account costs accrued on failed attempts).

Test Plan: Unit tests.

Reviewers: pboyapalli, ghorrell

Reviewed By: pboyapalli

Subscribers: pbohannon

Tags: #reflow

Maniphest Tasks: T37166

Differential Revision: https://phabricator.grailbio.com/D55461

fbshipit-source-id: 3db0b7f