[SYSINFRA-340] reflow/taskdb: Use same type for poolID when creating alloc rows
Summary:
When taskdb Alloc rows are created, pass poolID in the same type (`StringDigest`)
as that used when the Pool row is created.

This allows the taskdb implementation to define (and hide) the relationship between
`Pool` rows and `Alloc` rows.  For example, the fact that:
`Alloc.PoolID` == `Pool.ID` is now unknown and an irrelevant implementation detail
to the callers of `StartAlloc`.

Test Plan: Unit tests.

Reviewers: pboyapalli, ghorrell

Reviewed By: pboyapalli

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D59336

fbshipit-source-id: ed4bdb8