[SYSINFRA-72] reflow/sched: correct the StateWait retry logic in the scheduler
Summary:
Currently, the scheduler retries the same step if it encounters a retryable
error. In most cases this is the desired behavior, but for `StateWait` this
behavior is essentially a no-op.

A retry of the `StateWait` step just returns immediately because the `x.Wait`
call returns the same exec error which caused the initial failure (without
actually retrying anything). So, in this case we actually need to go back to
`StatePut` to re-submit the exec to the alloc and execute a real retry.

Additionally, we now retry a task up to `numExecTries` times total, whereas
before each step could be retried up to `numExecTries` times. This was done to
simplify the bookkeeping and avoid introducing cycles because there is no longer
a guarantee that the state only moves forward.

Test Plan:
Existing and newly added unit tests with 100% coverage of `ExecState.Next`
passed.

Reviewers: smahadevan, ghorrell, pbohannon

Reviewed By: ghorrell

Subscribers: dnicolaou

Tags: #reflow

Differential Revision: https://phabricator.grailbio.com/D55938

fbshipit-source-id: 08e9e80